Cms: fix news module

+++ Cms/V1/Module/News/Controller/Adminhtml/Index/Delete.php
@@ -1,30 +1,24 @@
 <?php
 
-namespace Cms\News\Controller\Adminhtml\News;
+namespace Cms\News\Controller\Adminhtml\Index;
 
 use Magento\Backend\App\Action;
 use Magento\Backend\App\Action\Context;
 use Magento\Framework\Controller\ResultFactory;
 use Magento\Framework\Controller\ResultInterface;
-use Magento\Ui\Component\MassAction\Filter;
 use Cms\News\Api\NewsRepositoryInterface;
 use Cms\News\Model\ResourceModel\News\CollectionFactory;
-use Cms\News\Service\NewsRepository;
 
-class MassDelete extends Action
+class Delete extends Action
 {
 
     const ADMIN_RESOURCE = 'Cms_News::delete';
-
-    private Filter $filter;
-
     private CollectionFactory $collectionFactory;
 
     private NewsRepositoryInterface $newsRepository;
-    public function __construct(Context $context, Filter $filter, CollectionFactory $collectionFactory, NewsRepositoryInterface $newsRepository)
+    public function __construct(Context $context, CollectionFactory $collectionFactory, NewsRepositoryInterface $newsRepository)
     {
         parent::__construct($context);
-        $this->filter = $filter;
         $this->collectionFactory = $collectionFactory;
         $this->newsRepository = $newsRepository;
     }
@@ -32,21 +26,24 @@ class MassDelete extends Action
     public function execute() : ResultInterface
     {
 
+        $newsId = $this->getRequest()->getParam('news_id', 0);
+
+        $result = $this->resultFactory->create(ResultFactory::TYPE_REDIRECT);
+
+        if (!$newsId) {
+            $this->messageManager->addWarningMessage(__('News not found.'));
+            return $result->setPath('news/news/index');
+        }
+
         try {
-            $collection = $this->filter->getCollection($this->collectionFactory->create());
-            $collectionSize = $collection->getSize();
 
+            $news = $this->newsRepository->getNewsById($newsId);
+            $this->newsRepository->delete($news);
 
-            foreach ($collection as $news) {
-                $this->newsRepository->delete($news);
-            }
-            $this->messageManager->addSuccessMessage(__('A total of %1 record(s) have been deleted.', $collectionSize));
+            $this->messageManager->addSuccessMessage(__('News has been deleted.'));
         } catch (\Exception $e) {
             $this->messageManager->addErrorMessage(__('There was an error while deleting the news!'));
         }
-
-
-        $result = $this->resultFactory->create(ResultFactory::TYPE_REDIRECT);
 
         return $result->setPath('news/news/index');
     }
